pyproject.toml
[tool.poetry]
name = "referrerlink"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.28.0"
fastapi = "^0.110.0"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
email-validator = "^2.1.1"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
cryptography = "^42.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
